[cpu_backend] Extract shared AVX2 helpers into internal header
Move the file-local AVX2 helper primitives of avx2_impl.cpp (exp256_ps,
rcp_ps, hsum_avx, swiglu/gelu polynomial approximations and the exp2
lookup table) into a new avx2_internal.h under namespace
nntrainer::avx2::internal, so that upcoming FP16 x86 kernels can reuse
the same primitives without forward declarations or code duplication.

The helpers are defined as plain inline functions and the gelu/erf
polynomial coefficients are constexpr constants scoped to the namespace.
The compiler shim macros (_nnt_CC_VECTORCALL, _nnt_ATTR_FLATTEN,
_nnt_ATTR_ALWAYS_INLINE and the __has_* fallbacks) are #undef'ed at the
end of the header, so it leaks no macros into including translation
units. avx2_impl.cpp reaches the primitives through explicit
using-declarations rather than a blanket using-directive, keeping its
dependencies visible.

Unreferenced code paths are dropped along the way: the scalar
approx_exp_exp2_lookup template, a commented-out low-precision variant
of exp256_ps, the compat_bit_cast/concept17_BinaryCastable helpers that
removal leaves without users, and the unused UNLIKELY and
_nnt_ATTR_NOINLINE macros.

The swiglu and gelu primitives were previously exercised only by an
Android-only test binary, so swiglu/tanh_gelu_v2/gelu_v2 accuracy tests
against the scalar fallback references are added to
unittest_nntrainer_cpu_backend to pin the moved code down on x86 CI. A
note in the x86 meson.build records that avx2_internal.h is intentionally
not installed with nntrainer_headers.

Also adds missing doxygen briefs on the q4_0 wrappers and rewrites a
plain multi-line comment in compute_fp16vcache_fp32_transposed to //
form, which the doxygen tag checker rejects.

Generated code is verified unchanged: compiling avx2_impl.cpp on this
commit and on main with identical flags yields the same symbol set and
instruction stream, differing only in the source line numbers and string
addresses that assert() embeds into its failure messages.
